Basic comment editing
Routes can only be used when there's a single {{outlet}} so we use
a @tracked conditional to implement an in-place-edit.

We use the built in <Textarea> component so that we don't have to 
manually handle input events that are to update ember's local model
state.

We have to override the comment adapter URL's to make it understand
Rails' nested URL scheme. By passing in @post={this.model} to the
Comment component we can set it on the comment via
this.args.comment.post = this.args.post before saving it. This makes it
available to the adapter via snapshot.belongsTo("post", { id: true }).

Next is cleaning up styling, see if we can use built-in input
components for Post too and add acceptance tests cases (possibly
cleaning up the mirage url resolve functions).